feat: StoryBook ProfileImage
